Reducing precision of observation causes 3x speedup

+++ mfec/agent.py
@@ -74,7 +74,7 @@ class MFECAgent:
         self.time += 1
 
         # Preprocess and project observation to state
-        self.state = np.dot(self.projection, np.asarray(observation).flatten())
+        self.state = np.dot(self.projection, np.asarray(observation, dtype=np.float32).flatten())
         # self.state = observation
 
         # Exploration
